Refactor: Extract year start and end epochms
