+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/endofyear/EndOfYearManagerImpl.kt
@@ -3,14 +3,10 @@ package au.com.shiftyjelly.pocketcasts.repositories.endofyear
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
-import au.com.shiftyjelly.pocketcasts.utils.DateUtil
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
-import timber.log.Timber
-import java.time.DateTimeException
-import java.time.LocalDate
 import javax.inject.Inject
 import kotlin.coroutines.CoroutineContext
 
@@ -25,20 +21,8 @@ class EndOfYearManagerImpl @Inject constructor(
         return podcastManager.findRandomPodcasts()
     }
 
-    override fun getTotalListeningTimeInSecsForYear(year: Int): Flow<Long?> {
-        return try {
-            val date = LocalDate.of(year, 1, 1).atStartOfDay()
-            val fromEpochTimeInMs = DateUtil.toEpochMillis(date)
-            val toEpochTimeInMs = DateUtil.toEpochMillis(date.plusYears(1))
-
-            if (fromEpochTimeInMs != null && toEpochTimeInMs != null) {
-                episodeManager.calculateListeningTime(fromEpochTimeInMs, toEpochTimeInMs)
-            } else {
-                flowOf(null)
-            }
-        } catch (e: DateTimeException) {
-            Timber.e(e)
-            flowOf(null)
-        }
-    }
+    override fun getTotalListeningTimeInSecsForYear(year: Int) =
+        getYearStartAndEndEpochMs(year)?.let {
+            episodeManager.calculateListeningTime(it.start, it.end)
+        } ?: flowOf(null)
 }
